testes dos validadores e da atualização de caminhão finalizados

+++ Models/Caminhoes/Caminhao.cs
@@ -5,7 +5,7 @@ namespace Models.Caminhoes
     public class Caminhao : BaseModel
     {
         public ModeloCaminhao Modelo { get; set; }
-        public short AnoFabricacao { get; set; }
-        public short AnoModelo { get; set; }
+        public int AnoFabricacao { get; set; }
+        public int AnoModelo { get; set; }
     }
 }
